[infra] Update repo page viewmodels dir structure
