+++ src/FluentHub/App.xaml.cs
@@ -92,11 +92,11 @@ namespace FluentHub
                 .AddTransient<ViewModels.AppSettings.AppearanceViewModel>()
                 .AddTransient<ViewModels.Home.NotificationsViewModel>()
                 .AddTransient<ViewModels.Home.ActivitiesViewModel>()
-                .AddTransient<ViewModels.Repositories.Layouts.DetailsLayoutViewModel>()
-                .AddTransient<ViewModels.Repositories.IssueViewModel>()
-                .AddTransient<ViewModels.Repositories.IssuesViewModel>()
-                .AddTransient<ViewModels.Repositories.PullRequestViewModel>()
-                .AddTransient<ViewModels.Repositories.PullRequestsViewModel>()
+                .AddTransient<ViewModels.Repositories.Codes.Layouts.DetailsLayoutViewModel>()
+                .AddTransient<ViewModels.Repositories.Issues.IssueViewModel>()
+                .AddTransient<ViewModels.Repositories.Issues.IssuesViewModel>()
+                .AddTransient<ViewModels.Repositories.PullRequests.PullRequestViewModel>()
+                .AddTransient<ViewModels.Repositories.PullRequests.PullRequestsViewModel>()
                 .AddTransient<ViewModels.UserControls.Blocks.FileContentBlockViewModel>()
                 .AddTransient<ViewModels.UserControls.Blocks.FileNavigationBlockViewModel>()
                 .AddTransient<ViewModels.UserControls.Blocks.ReadmeContentBlockViewModel>()
