+++ src/FluentHub/FluentHub.csproj
@@ -246,15 +246,15 @@
     <Compile Include="ViewModels\Organizations\OverviewViewModel.cs" />
     <Compile Include="ViewModels\Organizations\ProfileViewModel.cs" />
     <Compile Include="ViewModels\Organizations\RepositoriesViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\CodePageViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\Codes\CodePageViewModel.cs" />
     <Compile Include="ViewModels\Repositories\RepoContextViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\IssuesViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\IssueViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\Layouts\DetailsLayoutViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\Layouts\TreeLayoutViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\Issues\IssuesViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\Issues\IssueViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\Codes\Layouts\DetailsLayoutViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\Codes\Layouts\TreeLayoutViewModel.cs" />
     <Compile Include="ViewModels\Repositories\OverviewViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\PullRequestsViewModel.cs" />
-    <Compile Include="ViewModels\Repositories\PullRequestViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\PullRequests\PullRequestsViewModel.cs" />
+    <Compile Include="ViewModels\Repositories\PullRequests\PullRequestViewModel.cs" />
     <Compile Include="ViewModels\SettingsManager.cs" />
     <Compile Include="ViewModels\SettingsViewModel.cs" />
     <Compile Include="ViewModels\UserControls\Blocks\ActivityBlockViewModel.cs" />
