+++ src/FluentHub/ViewModels/Repositories/Codes/Layouts/TreeLayoutViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FluentHub.ViewModels.Repositories.Layouts
+namespace FluentHub.ViewModels.Repositories.Codes.Layouts
 {
     public class TreeLayoutViewModel
     {
