+++ src/FluentHub/ViewModels/Repositories/Issues/IssueViewModel.cs
@@ -4,7 +4,6 @@ using FluentHub.Models;
 using FluentHub.Octokit.Queries.Repositories;
 using FluentHub.UserControls.Blocks;
 using FluentHub.ViewModels.UserControls.Blocks;
-using Humanizer;
 using Microsoft.Toolkit.Mvvm.ComponentModel;
 using Microsoft.Toolkit.Mvvm.Input;
 using Microsoft.Toolkit.Mvvm.Messaging;
@@ -13,61 +12,64 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Threading;
 using System.Threading.Tasks;
+using Windows.UI.Xaml;
+using Windows.UI.Xaml.Controls;
 
-namespace FluentHub.ViewModels.Repositories
+namespace FluentHub.ViewModels.Repositories.Issues
 {
-    public class PullRequestViewModel : ObservableObject
+    public class IssueViewModel : ObservableObject
     {
         #region constructor
-        public PullRequestViewModel(IMessenger messenger = null, ILogger logger = null)
+        public IssueViewModel(IMessenger messenger = null, ILogger logger = null)
         {
             _messenger = messenger;
             _logger = logger;
 
             CommentBlocks = new();
-            RefreshPullRequestPageCommand = new AsyncRelayCommand<PullRequest>(RefreshPullRequestPageAsync);
+            RefreshIssuePageCommand = new AsyncRelayCommand<Issue>(RefreshIssuePageAsync);
         }
         #endregion
 
         #region fields
         private readonly IMessenger _messenger;
         private readonly ILogger _logger;
-        private PullRequest pullItem;
+        private IssueCommentBlockViewModel _commentViewModel;
+        private Issue _issueItem;
         #endregion
 
         #region properties
-        public PullRequest PullItem { get => pullItem; private set => SetProperty(ref pullItem, value); }
+        public Issue IssueItem { get => _issueItem; private set => SetProperty(ref _issueItem, value); }
+        public IssueCommentBlockViewModel CommentViewModel { get => _commentViewModel; private set => SetProperty(ref _commentViewModel, value); }
         public ObservableCollection<IssueCommentBlock> CommentBlocks;
-        public IAsyncRelayCommand RefreshPullRequestPageCommand { get; }
+        public IAsyncRelayCommand RefreshIssuePageCommand { get; }
         #endregion
 
         #region methods
-        private async Task RefreshPullRequestPageAsync(PullRequest pull)
+        private async Task RefreshIssuePageAsync(Issue issue)
         {
             try
             {
-                PullItem = pull;
+                IssueItem = issue;
 
-                PullRequestEventQueries queries = new();
-                var prBoddy = await queries.GetBodyAsync(PullItem.OwnerLogin, PullItem.Name, PullItem.Number);
+                IssueEventQueries queries = new();
+                var issueBody = await queries.GetBodyAsync(issue.OwnerLogin, issue.Name, issue.Number);
 
-                var bodyCommentViewModel = new IssueCommentBlockViewModel()
-                {
-                    IssueComment = prBoddy,
+                var bodyCommentViewModel = new IssueCommentBlockViewModel() {
+                    IssueComment = issueBody,
                 };
 
                 var bodyCommentBlock = new IssueCommentBlock() { PropertyViewModel = bodyCommentViewModel };
 
                 CommentBlocks.Add(bodyCommentBlock);
 
-                // Now only available PR comments
-                var prComments = await queries.GetAllAsync(PullItem.OwnerLogin, PullItem.Name, PullItem.Number);
+                // Now only available issue comments
+                var issueComments = await queries.GetAllAsync(issue.OwnerLogin, issue.Name, issue.Number);
 
-                foreach (var prCommentItem in prComments)
+                foreach(var issueCommentItem in issueComments)
                 {
                     var viewmodel = new IssueCommentBlockViewModel()
                     {
-                        IssueComment = prCommentItem,
+                        IssueComment = issueCommentItem,
                     };
 
                     var commentBlock = new IssueCommentBlock()
@@ -80,7 +82,7 @@ namespace FluentHub.ViewModels.Repositories
             }
             catch (Exception ex)
             {
-                _logger?.Error("RefreshPullRequestPageAsync", ex);
+                _logger?.Error("RefreshIssuePageAsync", ex);
                 if (_messenger != null)
                 {
                     UserNotificationMessage notification = new("Something went wrong", ex.Message, UserNotificationType.Error);
