+++ src/FluentHub/ViewModels/Repositories/PullRequests/PullRequestViewModel.cs
@@ -4,6 +4,7 @@ using FluentHub.Models;
 using FluentHub.Octokit.Queries.Repositories;
 using FluentHub.UserControls.Blocks;
 using FluentHub.ViewModels.UserControls.Blocks;
+using Humanizer;
 using Microsoft.Toolkit.Mvvm.ComponentModel;
 using Microsoft.Toolkit.Mvvm.Input;
 using Microsoft.Toolkit.Mvvm.Messaging;
@@ -12,64 +13,61 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Threading;
 using System.Threading.Tasks;
-using Windows.UI.Xaml;
-using Windows.UI.Xaml.Controls;
 
-namespace FluentHub.ViewModels.Repositories
+namespace FluentHub.ViewModels.Repositories.PullRequests
 {
-    public class IssueViewModel : ObservableObject
+    public class PullRequestViewModel : ObservableObject
     {
         #region constructor
-        public IssueViewModel(IMessenger messenger = null, ILogger logger = null)
+        public PullRequestViewModel(IMessenger messenger = null, ILogger logger = null)
         {
             _messenger = messenger;
             _logger = logger;
 
             CommentBlocks = new();
-            RefreshIssuePageCommand = new AsyncRelayCommand<Issue>(RefreshIssuePageAsync);
+            RefreshPullRequestPageCommand = new AsyncRelayCommand<PullRequest>(RefreshPullRequestPageAsync);
         }
         #endregion
 
         #region fields
         private readonly IMessenger _messenger;
         private readonly ILogger _logger;
-        private IssueCommentBlockViewModel _commentViewModel;
-        private Issue _issueItem;
+        private PullRequest pullItem;
         #endregion
 
         #region properties
-        public Issue IssueItem { get => _issueItem; private set => SetProperty(ref _issueItem, value); }
-        public IssueCommentBlockViewModel CommentViewModel { get => _commentViewModel; private set => SetProperty(ref _commentViewModel, value); }
+        public PullRequest PullItem { get => pullItem; private set => SetProperty(ref pullItem, value); }
         public ObservableCollection<IssueCommentBlock> CommentBlocks;
-        public IAsyncRelayCommand RefreshIssuePageCommand { get; }
+        public IAsyncRelayCommand RefreshPullRequestPageCommand { get; }
         #endregion
 
         #region methods
-        private async Task RefreshIssuePageAsync(Issue issue)
+        private async Task RefreshPullRequestPageAsync(PullRequest pull)
         {
             try
             {
-                IssueItem = issue;
+                PullItem = pull;
 
-                IssueEventQueries queries = new();
-                var issueBody = await queries.GetBodyAsync(issue.OwnerLogin, issue.Name, issue.Number);
+                PullRequestEventQueries queries = new();
+                var prBoddy = await queries.GetBodyAsync(PullItem.OwnerLogin, PullItem.Name, PullItem.Number);
 
-                var bodyCommentViewModel = new IssueCommentBlockViewModel() {
-                    IssueComment = issueBody,
+                var bodyCommentViewModel = new IssueCommentBlockViewModel()
+                {
+                    IssueComment = prBoddy,
                 };
 
                 var bodyCommentBlock = new IssueCommentBlock() { PropertyViewModel = bodyCommentViewModel };
 
                 CommentBlocks.Add(bodyCommentBlock);
 
-                // Now only available issue comments
-                var issueComments = await queries.GetAllAsync(issue.OwnerLogin, issue.Name, issue.Number);
+                // Now only available PR comments
+                var prComments = await queries.GetAllAsync(PullItem.OwnerLogin, PullItem.Name, PullItem.Number);
 
-                foreach(var issueCommentItem in issueComments)
+                foreach (var prCommentItem in prComments)
                 {
                     var viewmodel = new IssueCommentBlockViewModel()
                     {
-                        IssueComment = issueCommentItem,
+                        IssueComment = prCommentItem,
                     };
 
                     var commentBlock = new IssueCommentBlock()
@@ -82,7 +80,7 @@ namespace FluentHub.ViewModels.Repositories
             }
             catch (Exception ex)
             {
-                _logger?.Error("RefreshIssuePageAsync", ex);
+                _logger?.Error("RefreshPullRequestPageAsync", ex);
                 if (_messenger != null)
                 {
                     UserNotificationMessage notification = new("Something went wrong", ex.Message, UserNotificationType.Error);
