+++ src/FluentHub/ViewModels/Repositories/PullRequests/PullRequestsViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using FluentHub.Backend;
-using FluentHub.Models;
 using FluentHub.Octokit.Models;
+using FluentHub.Models;
 using FluentHub.Octokit.Queries.Repositories;
 using FluentHub.ViewModels.UserControls.ButtonBlocks;
 using Humanizer;
@@ -13,62 +13,62 @@ using System.Collections.ObjectModel;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace FluentHub.ViewModels.Repositories
+namespace FluentHub.ViewModels.Repositories.PullRequests
 {
-    public class IssuesViewModel : ObservableObject
+    public class PullRequestsViewModel : ObservableObject
     {
         #region constructor
-        public IssuesViewModel(IMessenger messenger = null, ILogger logger = null)
+        public PullRequestsViewModel(IMessenger messenger = null, ILogger logger = null)
         {
             _messenger = messenger;
             _logger = logger;
             _messenger = messenger;
-            _issueItems = new();
-            IssueItems = new(_issueItems);
+            _pullRequests = new();
+            PullItems = new(_pullRequests);
 
-            RefreshIssuesPageCommand = new AsyncRelayCommand<string>(RefreshIssuesPageAsync);
+            RefreshPullRequestsPageCommand = new AsyncRelayCommand<string>(RefreshPullRequestsPageAsync);
         }
         #endregion
 
         #region fields
-        private readonly ILogger _logger;
         private readonly IMessenger _messenger;
-        private readonly ObservableCollection<IssueButtonBlockViewModel> _issueItems;
+        private readonly ILogger _logger;
+        private readonly ObservableCollection<PullButtonBlockViewModel> _pullRequests;
         #endregion
 
         #region properties
-        public ReadOnlyObservableCollection<IssueButtonBlockViewModel> IssueItems { get; }
-        public IAsyncRelayCommand RefreshIssuesPageCommand { get; }
+        public ReadOnlyObservableCollection<PullButtonBlockViewModel> PullItems { get; }
+        public IAsyncRelayCommand RefreshPullRequestsPageCommand { get; }
         #endregion
 
         #region methods
-        private async Task RefreshIssuesPageAsync(string nameWithOwner, CancellationToken token)
+        private async Task RefreshPullRequestsPageAsync(string nameWithOwner, CancellationToken token)
         {
             try
             {
-                IssueQueries queries = new();
-                List<Issue> items = await queries.GetAllAsync(nameWithOwner.Split("/")[1], nameWithOwner.Split("/")[0]);
+                PullRequestQueries queries = new();
+                List<PullRequest> items = await queries.GetAllAsync(nameWithOwner.Split("/")[1], nameWithOwner.Split("/")[0]);
 
                 if (items == null) return;
 
-                _issueItems.Clear();
+                _pullRequests.Clear();
 
                 foreach (var item in items)
                 {
-                    IssueButtonBlockViewModel viewModel = new()
+                    PullButtonBlockViewModel viewModel = new()
                     {
-                        IssueItem = item,
-                        NameWithOwner = item.OwnerLogin + "/" + item.Name + " #" + item.Number,
+                        PullItem = item,
+                        NameWithOwner = $"{item.OwnerLogin} / {item.Name} #{item.Number}",
                         UpdatedAtHumanized = item.UpdatedAt.Humanize()
                     };
 
-                    _issueItems.Add(viewModel);
+                    _pullRequests.Add(viewModel);
                 }
             }
             catch (OperationCanceledException) { }
             catch (Exception ex)
             {
-                _logger?.Error("RefreshIssuesPageAsync", ex);
+                _logger?.Error("RefreshPullRequestsAsync", ex);
                 if (_messenger != null)
                 {
                     UserNotificationMessage notification = new("Something went wrong", ex.Message, UserNotificationType.Error);
