+++ src/FluentHub/Views/Repositories/Codes/CodePage.xaml
@@ -1,13 +1,13 @@
 ﻿<Page
-    x:Class="FluentHub.Views.Repositories.CodePage"
+    x:Class="FluentHub.Views.Repositories.Codes.CodePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:FluentHub.Views.Repositories"
+    xmlns:local="using:FluentHub.Views.Repositories.Codes"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     xmlns:uc="using:FluentHub.UserControls"
-    xmlns:vm="using:FluentHub.ViewModels.Repositories"
+    xmlns:vm="using:FluentHub.ViewModels.Repositories.Codes"
     mc:Ignorable="d">
 
     <Page.DataContext>
