+++ src/FluentHub/Views/Repositories/Codes/CodePage.xaml.cs
@@ -1,21 +1,18 @@
-﻿using FluentHub.ViewModels.Repositories;
+﻿using FluentHub.Models.Items;
+using FluentHub.Services;
+using FluentHub.Services.Navigation;
+using FluentHub.ViewModels;
+using FluentHub.ViewModels.Repositories;
+using FluentHub.ViewModels.Repositories.Codes;
+using Microsoft.Extensions.DependencyInjection;
 using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
 using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
-using Windows.UI.Xaml.Media.Imaging;
 using Windows.UI.Xaml.Navigation;
+using Windows.UI.Xaml.Media.Imaging;
 
-namespace FluentHub.Views.Repositories
+namespace FluentHub.Views.Repositories.Codes
 {
     public sealed partial class CodePage : Page
     {
