+++ src/FluentHub/Views/Repositories/Codes/Layouts/TreeLayoutView.xaml
@@ -1,9 +1,9 @@
 ﻿<Page
-    x:Class="FluentHub.Views.Repositories.Layouts.TreeLayoutView"
+    x:Class="FluentHub.Views.Repositories.Codes.Layouts.TreeLayoutView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:FluentHub.Views.Repositories.Layouts"
+    xmlns:local="using:FluentHub.Views.Repositories.Codes.Layouts"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
