+++ src/FluentHub/Views/Repositories/Codes/ReleasesPage.xaml.cs
@@ -13,13 +13,8 @@ using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 
-// 空白ページの項目テンプレートについては、https://go.microsoft.com/fwlink/?LinkId=234238 を参照してください
-
 namespace FluentHub.Views.Repositories.Codes
 {
-    /// <summary>
-    /// それ自体で使用できる空白ページまたはフレーム内に移動できる空白ページ。
-    /// </summary>
     public sealed partial class ReleasesPage : Page
     {
         public ReleasesPage()
