+++ src/FluentHub/Views/Repositories/Issues/IssuePage.xaml
@@ -1,12 +1,11 @@
 ﻿<Page
-    x:Class="FluentHub.Views.Repositories.IssuePage"
+    x:Class="FluentHub.Views.Repositories.Issues.IssuePage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:block="using:FluentHub.UserControls.ButtonBlocks"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:item="using:FluentHub.Models.Items"
     xmlns:label="using:FluentHub.UserControls.Labels"
-    xmlns:local="using:FluentHub.Views.Repositories"
+    xmlns:local="using:FluentHub.Views.Repositories.Issues"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     xmlns:toolkit="using:Microsoft.Toolkit.Uwp.UI.Controls"
