+++ src/FluentHub/Views/Repositories/PullRequests/PullRequestPage.xaml
@@ -1,10 +1,10 @@
 ﻿<Page
-    x:Class="FluentHub.Views.Repositories.PullRequestPage"
+    x:Class="FluentHub.Views.Repositories.PullRequests.PullRequestPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:label="using:FluentHub.UserControls.Labels"
-    xmlns:local="using:FluentHub.Views.Repositories"
+    xmlns:local="using:FluentHub.Views.Repositories.PullRequests"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:muxc="using:Microsoft.UI.Xaml.Controls"
     xmlns:toolkit="using:Microsoft.Toolkit.Uwp.UI.Controls"
